refactor(server): split routes/parse.rs into a parse/ facade module
parse.rs (1342 LOC) carried six independent route concerns. Convert it into
a parse/ directory behind a mod.rs facade:

  mod.rs           ParseQuery + extract_file + `pub use` of every handler
  cache_keys.rs    cache-key builders + symbolic cache helpers + tests
  json.rs          parse_full / parse_stream / parse_metadata
  parquet.rs       parse_parquet / parse_parquet_optimized + headers
  parquet_stream.rs parse_parquet_stream + event type
  fetch.rs         get_data_model / get_symbolic / check_cache / get_cached_geometry

Pure code motion. Handlers + ParseQuery are re-exported from mod.rs so all 10
`routes::parse::*` references in main.rs resolve unchanged; main.rs and
routes/mod.rs are untouched. Cache-key format! literals were moved verbatim
(no re-hashing). Cross-submodule items gain pub(crate); the rest stay private.
cargo test -p ifc-lite-server green (21); clippy -D warnings clean.
